relay buy storage events from evm to jackal

+++ forge/src/StorageDrawer.sol
@@ -4,12 +4,11 @@ pragma solidity ^0.8.26;
 import {JackalInterface} from "./JackalInterface.sol";
 
 contract StorageDrawer {
-
     JackalInterface internal jackalBridge;
 
     mapping(address => string[]) public cabinet;
 
-    constructor(address _jackalAddress){
+    constructor(address _jackalAddress) {
         jackalBridge = JackalInterface(_jackalAddress);
     }
 
@@ -18,12 +17,12 @@ contract StorageDrawer {
         cabinet[msg.sender].push(merkle);
     }
 
-    function fileAddress(address _addr, uint _index) public view returns (string memory){
+    function fileAddress(address _addr, uint256 _index) public view returns (string memory) {
         require(_index < cabinet[_addr].length);
         return cabinet[_addr][_index];
     }
 
-    function fileCount(address _addr) public view returns (uint){
+    function fileCount(address _addr) public view returns (uint256) {
         return cabinet[_addr].length;
     }
 }
